+++ glyph-brush-layout/CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 * Re-export `rusttype::point`.
+* Fix `bounds_rect` implementation for some `f32::INFINITY` cases.
 
 # 0.1
 * Initial release.